+++ CHANGELOG.md
@@ -5,6 +5,8 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+- Added `lookup` built-in;
+
 ## 0.16.1
 
 ### Fixes